separate auth files and improve doc coments
