+++ Sources/MusicLibrary.swift
@@ -102,18 +102,53 @@ public final class MusicLibrary: MusicLibraryProtocol {
         self.service = service
     }
 
-    // MARK: - General calls
+    // MARK: - Authorization
 
+    /// Requests music library access authorization from the user.
+    ///
+    /// Presents the system authorization prompt if the user hasn't yet decided.
+    /// If authorization is already determined, returns the current status without prompting.
+    /// Delegates to the underlying ``AuthorizationManagerProtocol`` implementation.
+    ///
+    /// - Returns: The authorization status after the request
+    /// - Throws: ``AuthorizationManagerError/unauthorized(_:)`` if the request fails or is denied
     @discardableResult
     public func requestAuthorization() async throws -> MPMediaLibraryAuthorizationStatus {
         try await auth.authorize()
     }
+    
+    // MARK: - General Media Queries
 
+    /// Fetches all media items of a specific type.
+    ///
+    /// Retrieves all media items matching the specified type from the device's music library.
+    /// Automatically checks authorization status before making the query.
+    ///
+    /// - Parameters:
+    ///   - type: The type of media to fetch (typically `.music`)
+    ///   - groupingType: How to group the returned items (`.title`, `.album`, `.artist`, etc.)
+    /// - Returns: Array of all media items matching the specified type
+    /// - Throws: ``AuthorizationManagerError/unauthorized(_:)`` if music library access is not authorized,
+    ///   or ``MusicLibraryServiceError/noItemsFound`` if no items are found
     public func fetchAll(_ type: MPMediaType, groupingType: MPMediaGrouping) async throws -> [MPMediaItem] {
         try await checkIfAuthorized()
         return try await service.fetchAll(type, groupingType: groupingType)
     }
 
+    /// Fetches media items of a specific type matching a predicate.
+    ///
+    /// Queries the music library for items matching the provided type and predicate.
+    /// This is the most flexible method, allowing you to fetch any media type (music, podcasts, audiobooks)
+    /// with custom filtering and grouping. Automatically checks authorization before querying.
+    ///
+    /// - Parameters:
+    ///   - type: The type of media to fetch (`.music`, `.podcast`, etc.)
+    ///   - predicate: The predicate to filter items (e.g., `.artist("Taylor Swift")`, `.genre("Rock")`)
+    ///   - comparisonType: How to compare the predicate value (`.equalTo`, `.contains`, etc.)
+    ///   - groupingType: How to group the returned items
+    /// - Returns: Array of media items matching the criteria
+    /// - Throws: ``AuthorizationManagerError/unauthorized(_:)`` if music library access is not authorized,
+    ///   or ``MusicLibraryServiceError/noItemFound(_:)`` if no matching items are found
     public func mediaItems(
         ofType type: MPMediaType,
         matching predicate: MediaItemPredicateInfo,
@@ -124,6 +159,20 @@ public final class MusicLibrary: MusicLibraryProtocol {
         return try await service.fetch(type, with: predicate, comparisonType: comparisonType, groupingType: groupingType)
     }
 
+    /// Fetches media item collections of a specific type matching a predicate.
+    ///
+    /// Queries the music library for item collections matching the provided type and predicate.
+    /// Returns grouped results (e.g., albums, artists) rather than individual items.
+    /// Automatically checks authorization before querying.
+    ///
+    /// - Parameters:
+    ///   - type: The type of media to fetch (`.music`, `.podcast`, etc.)
+    ///   - predicate: The predicate to filter collections (e.g., `.albumArtist("The Beatles")`)
+    ///   - comparisonType: How to compare the predicate value (`.equalTo`, `.contains`, etc.)
+    ///   - groupingType: How to group the returned collections (typically `.album` or `.albumArtist`)
+    /// - Returns: Array of media item collections matching the criteria
+    /// - Throws: ``AuthorizationManagerError/unauthorized(_:)`` if music library access is not authorized,
+    ///   or ``MusicLibraryServiceError/noCollectionFound(_:)`` if no matching collections are found
     public func mediaItemCollections(
         ofType type: MPMediaType,
         matching predicate: MediaItemPredicateInfo,
@@ -174,6 +223,29 @@ public final class MusicLibrary: MusicLibraryProtocol {
         return try await mediaItems(ofType: .music, matching: predicate, comparisonType, groupingType: .title)
     }
 
+    /// Fetches album collections matching a predicate.
+    ///
+    /// Convenience method for fetching music albums grouped by the specified grouping type.
+    /// This is a specialized version of ``mediaItemCollections(ofType:matching:_:groupingType:)``
+    /// that's pre-configured for music albums.
+    ///
+    /// - Parameters:
+    ///   - predicate: The predicate to filter albums (e.g., `.albumArtist("The Beatles")`, `.genre("Rock")`)
+    ///   - comparisonType: How to compare the predicate value (`.equalTo`, `.contains`, etc.)
+    ///   - groupingType: How to group the returned collections (typically `.album` or `.albumArtist`)
+    /// - Returns: Array of album collections matching the criteria
+    /// - Throws: ``AuthorizationManagerError/unauthorized(_:)`` if music library access is not authorized,
+    ///   or ``MusicLibraryServiceError/noCollectionFound(_:)`` if no matching albums are found
+    ///
+    /// ## Example
+    /// ```swift
+    /// let library = MusicLibrary()
+    /// let beatlesAlbums = try await library.albums(
+    ///     matching: .albumArtist("The Beatles"),
+    ///     .equalTo,
+    ///     groupingType: .album
+    /// )
+    /// ```
     public func albums(
         matching predicate: MediaItemPredicateInfo,
         _ comparisonType: MPMediaPredicateComparison,
